Add logic for search area
Allow rover to detect when it is in the search area. Once it does,
rotate to the left until it senses the target object. Then, move up to
the target object and pick it up. Rotate left until backside is facing
the left side of the area (drop zone). Move into drop zone and put down
target object.